# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SANOVIM — Marketing Médico com IA API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content/generate-post:
    post:
      operationId: generatePost
      tags: [content]
      summary: Generate Instagram post caption with Claude AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePostRequest"
      responses:
        "200":
          description: Generated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContent"
  /content/generate-reels-script:
    post:
      operationId: generateReelsScript
      tags: [content]
      summary: Generate Reels script (30-60s)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateReelsScriptRequest"
      responses:
        "200":
          description: Generated script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContent"
  /content/trends:
    get:
      operationId: getTrends
      tags: [content]
      summary: Top 5 trends in medical niche via Serper
      parameters:
        - name: account
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/AccountName"
        - name: specialty
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Medical trends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendsResponse"
  /content/innovations:
    get:
      operationId: getInnovations
      tags: [content]
      summary: Latest specialty innovations and news
      parameters:
        - name: specialty
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Innovations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InnovationsResponse"
  /content/history:
    get:
      operationId: getContentHistory
      tags: [content]
      summary: Generated content history
      parameters:
        - name: account
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AccountName"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Content history list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentHistoryItem"

  # ── Images ────────────────────────────────────────────────────────────
  /images/generate:
    post:
      operationId: generateImage
      tags: [images]
      summary: Generate image from text prompt via Runware
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateImageRequest"
      responses:
        "200":
          description: Generated image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"
  /images/create-card:
    post:
      operationId: createProfessionalCard
      tags: [images]
      summary: Create professional medical card (1080x1350)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateCardRequest"
      responses:
        "200":
          description: Card image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"
  /images/history:
    get:
      operationId: getImageHistory
      tags: [images]
      summary: Generated images history
      parameters:
        - name: account
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AccountName"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Images history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImageHistoryItem"

  # ── Metrics ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AccountName:
      type: string
      enum: [drdaniel, angelica, loysby]
...
    GeneratePostRequest:
      type: object
      properties:
        account:
          $ref: "#/components/schemas/AccountName"
        topic:
          type: string
        specialty:
          type: string
        tone:
          type: string
          enum: [educational, inspirational, promotional, awareness]
          default: educational
        includeHashtags:
          type: boolean
          default: true
        cfmCompliant:
          type: boolean
          default: true
      required: [account, topic]
    GenerateReelsScriptRequest:
      type: object
      properties:
        account:
          $ref: "#/components/schemas/AccountName"
        topic:
          type: string
        specialty:
          type: string
        durationSeconds:
          type: integer
          minimum: 30
          maximum: 60
          default: 45
      required: [account, topic]
    GeneratedContent:
      type: object
      properties:
        id:
          type: integer
        account:
          $ref: "#/components/schemas/AccountName"
        type:
          type: string
          enum: [post, reels_script]
        content:
          type: string
        hashtags:
          type: array
          items:
            type: string
        specialty:
          type: string
        cfmCompliant:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, account, type, content, createdAt]
    ContentHistoryItem:
      type: object
      properties:
        id:
          type: integer
        account:
          $ref: "#/components/schemas/AccountName"
        type:
          type: string
        topic:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, account, type, content, createdAt]
    TrendsResponse:
      type: object
      properties:
        account:
          $ref: "#/components/schemas/AccountName"
        specialty:
          type: string
        trends:
          type: array
          items:
            $ref: "#/components/schemas/TrendItem"
        fetchedAt:
          type: string
          format: date-time
      required: [trends, fetchedAt]
    TrendItem:
      type: object
      properties:
        rank:
          type: integer
        title:
          type: string
        description:
          type: string
        source:
          type: string
        url:
          type: string
      required: [rank, title, description]
    InnovationsResponse:
      type: object
      properties:
        specialty:
          type: string
        innovations:
          type: array
          items:
            $ref: "#/components/schemas/InnovationItem"
        fetchedAt:
          type: string
          format: date-time
      required: [innovations, fetchedAt]
    InnovationItem:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        source:
          type: string
        url:
          type: string
        date:
          type: string
      required: [title, summary]
    GenerateImageRequest:
      type: object
      properties:
        prompt:
          type: string
        account:
          $ref: "#/components/schemas/AccountName"
        style:
          type: string
          default: professional medical
        width:
          type: integer
          default: 1080
        height:
          type: integer
          default: 1350
      required: [prompt]
    CreateCardRequest:
      type: object
      properties:
        account:
          $ref: "#/components/schemas/AccountName"
        text:
          type: string
        subtext:
          type: string
        image:
          type: string
          format: binary
      required: [account, text]
    GeneratedImage:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        thumbnailUrl:
          type: string
        width:
          type: integer
        height:
          type: integer
        account:
          $ref: "#/components/schemas/AccountName"
        prompt:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, url, createdAt]
    ImageHistoryItem:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        account:
          $ref: "#/components/schemas/AccountName"
        type:
          type: string
          enum: [generated, card]
        prompt:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, url, createdAt]
...