# lib/api-zod/src/generated/types/accountName.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SANOVIM — Marketing Médico com IA API
 * OpenAPI spec version: 0.1.0
 */

export type AccountName = (typeof AccountName)[keyof typeof AccountName];

export const AccountName = {
  drdaniel: "drdaniel",
  angelica: "angelica",
  loysby: "loysby",
} as const;